Changing input parameters for read noise and gain from data models to simple numpy arrays.  Removed reference to jwst.reffile_utils from ramp fit processing code and moved it to ramp fit step code.
