+++ jwst/ramp_fitting/tests/test_ramp_fit_cases.py
@@ -25,16 +25,16 @@ def test_pix_0():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55.,
-                                            65., 75., 94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     # [data, dq, err, var_p, var_r]
@@ -60,16 +60,16 @@ def test_pix_1():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 4, 4, 0, 2, 2, 2, 2, 2, 2])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.8999999, 6, 1.046670, 0.02636364, 1.0691562]
@@ -90,16 +90,16 @@ def test_pix_2():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75., 94.,
-                                            95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 0, 4, 0, 4, 0, 4, 2, 2])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [0.84761256, 6, 0.42986465, 0.00659091, 0.1781927]
@@ -127,16 +127,16 @@ def test_pix_3():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 0, 0, 0, 0, 0, 0, 4, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.0746869, 4, 0.12186482, 0.00227273, 0.01257831]
@@ -164,16 +164,16 @@ def test_pix_4():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 1055., 1065., 1075.,
-                                            2594., 2595., 2605.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 1055., 1065., 1075., 2594., 2595., 2605.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 2, 2, 2, 2, 2, 2, 2, 2, 2])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.5, 2, 1.047105, 0.02727273, 1.0691562]
@@ -194,16 +194,16 @@ def test_pix_5():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 2055., 2065., 2075.,
-                                            2094., 2095., 2105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 2055., 2065., 2075., 2094., 2095., 2105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 0, 0, 4, 0, 0, 0, 0, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.076075, 4, 0.16134359, 0.00227273, 0.02375903]
@@ -231,16 +231,16 @@ def test_pix_6():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 375.,
-                                            394., 395., 405.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 375., 394., 395., 405.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 4, 4, 0, 0, 0, 0, 0, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [6.092052, 4, 0.14613187, 0.0025974, 0.01875712]
@@ -267,16 +267,16 @@ def test_pix_7():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 195., 205.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 195., 205.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 0, 0, 0, 0, 0, 0, 4, 4])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.0757396, 4, 0.12379601, 0.0025974, 0.01272805]
@@ -297,16 +297,16 @@ def test_pix_8():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 4, 0, 0, 0, 0, 0, 2, 2, 2])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.0101178, 6, 0.1848883, 0.00363636, 0.03054732]
@@ -328,16 +328,16 @@ def test_pix_9():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 4, 4, 0, 0, 0, 0, 4, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [0.9999994, 4, 0.22721863, 0.0030303, 0.048598]
@@ -366,16 +366,16 @@ def test_pix_10():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 4, 0, 0, 4, 0, 0, 0, 0])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1., 4, 0.21298744, 0.0025974, 0.04276625]
@@ -402,16 +402,16 @@ def test_pix_11():
     """
 
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([15., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 105.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [15., 25., 35., 54., 55., 65., 75., 94., 95., 105.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([0, 0, 2, 2, 2, 2, 2, 2, 2, 2])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1., 2, 1.042755, 0.01818182, 1.0691562]
@@ -434,8 +434,8 @@ def test_pix_12():
     ngroups = 2
     nints = 1
     ncols = 2
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
     ramp_model.data[0, :, 0, 0] = np.array([15., 59025.], dtype=np.float32)
@@ -444,8 +444,9 @@ def test_pix_12():
     ramp_model.groupdq[0, :, 0, 1] = np.array([2, 2])
 
     # call ramp_fit
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+
     # Set truth values for PRIMARY results for pixel 1:
     # slope, dq, err, var_p, var_r
     # slope = group1 / deltatime = 15 / 10 = 1.5
@@ -484,16 +485,16 @@ def test_miri_0():
     GROUPDQ is: [1, 0, 0, 0, 0, 0, 0, 0, 0, 1]
     """
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([8888., 25., 35., 54., 55.,
-                                            65., 75., 94., 95., 888.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [8888., 25., 35., 54., 55., 65., 75., 94., 95., 888.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([1, 0, 0, 0, 0, 0, 0, 0, 0, 1])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.025854, 0, 0.12379601, 0.0025974, 0.01272805]
@@ -514,16 +515,16 @@ def test_miri_1():
     GROUPDQ is: [5, 0, 0, 0, 0, 0, 0, 0, 0, 1]
     """
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([7777., 125., 135., 154., 165., 175.,
-                                            185., 204., 205., 777.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [7777., 125., 135., 154., 165., 175., 185., 204., 205., 777.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([5, 0, 0, 0, 0, 0, 0, 0, 0, 1])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.1996487, 0, 0.12379601, 0.0025974, 0.01272805]
@@ -544,16 +545,16 @@ def test_miri_2():
     GROUPDQ is: [5, 0, 0, 0, 0, 0, 0, 0, 0, 5]
     """
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([4444., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 444.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [4444., 25., 35., 54., 55., 65., 75., 94., 95., 444.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([5, 0, 0, 0, 0, 0, 0, 0, 0, 5])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.025854, 0, 0.12379601, 0.0025974, 0.01272805]
@@ -574,16 +575,16 @@ def test_miri_3():
     GROUPDQ is: [1, 0, 0, 0, 0, 0, 0, 0, 0, 5]
     """
     ngroups, nints, nrows, ncols, deltatime, gain, readnoise = set_scalars()
-    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(ngroups,
-                                                                                  nints, nrows, ncols, deltatime, gain, readnoise)
+    ramp_model, rnoise_model, gain_model, pixdq, groupdq, err = create_mod_arrays(
+        ngroups, nints, nrows, ncols, deltatime, gain, readnoise)
 
     # Populate pixel-specific SCI and GROUPDQ arrays
-    ramp_model.data[0, :, 0, 0] = np.array([6666., 25., 35., 54., 55., 65., 75.,
-                                            94., 95., 666.], dtype=np.float32)
+    ramp_model.data[0, :, 0, 0] = np.array(
+        [6666., 25., 35., 54., 55., 65., 75., 94., 95., 666.], dtype=np.float32)
     ramp_model.groupdq[0, :, 0, 0] = np.array([1, 0, 0, 0, 0, 0, 0, 0, 0, 5])
 
-    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(ramp_model, 1024 * 300., True,
-                                                      rnoise_model, gain_model, 'OLS', 'optimal', 'none')
+    new_mod, int_mod, opt_mod, gls_opt_mod = ramp_fit(
+        ramp_model, 1024 * 300., True, rnoise_model, gain_model, 'OLS', 'optimal', 'none')
 
     # Set truth values for PRIMARY results:
     p_true = [1.025854, 0, 0.12379601, 0.0025974, 0.01272805]
@@ -692,6 +693,8 @@ def create_mod_arrays(ngroups, nints, nrows, ncols, deltatime, gain, readnoise):
     ramp_model.meta.exposure.nframes = 1
     ramp_model.meta.exposure.groupgap = 0
     ramp_model.meta.exposure.drop_frames1 = 0
+
+    return ramp_model, read_noise, gain, pixdq, gdq, err
 
     # Create and populate gain model
     gain_model = GainModel(data=gain)
